requirements/local.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-quill-editor==0.1.40
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
sqlparse==0.5.0
tzdata==2024.1